pyproject.toml
[project]
name = "sales-assistant-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bson>=0.5.10",
    "fastapi[standard]>=0.121.2",
    "langchain>=1.0.7",
    "langchain-anthropic>=1.0.4",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.5",
    "langchain-google-genai>=3.0.3",
    "langchain-mcp-adapters>=0.1.13",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.3",
    "langgraph-supervisor>=0.0.30",
    "langmem>=0.0.30",
    "mcp>=1.21.1",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.3.0",
    "redis>=6.2.0",
    "redisvl>=0.3.9",
]
